+++ app/schemas/porutham_shares.py
@@ -74,7 +74,9 @@ class PoruthamShareViewData(BaseModel):
     girl_nakshatra: int = Field(alias="girlNakshatra")
     girl_nakshatra_name: str = Field(alias="girlNakshatraName")
     kutas: list[KutaResult]
-    total_score: int = Field(alias="totalScore")
+    # float since 2026-08-31: a madhyama contributes 0.5. Rendered via
+    # `format_porutham_total` so a whole score never reads "8.0/10".
+    total_score: float = Field(alias="totalScore")
     max_score: int = Field(alias="maxScore")
     percentage: float
     label: str
